Changed usage of ImageInterface<double> to ImageInterface<float>, since the viewer is unable to handle the higher precision.
git-svn-id: http://usg.lofar.org/svn/code/trunk@2476 e3a606d5-b720-0410-b6ad-cb58f335abf8

+++ src/CR-Tools/implement/Imaging/Skymapper.cc
@@ -310,11 +310,11 @@ namespace CR {  // Namespace CR -- begin
 
     /* Create paged image on disk */
 // #ifdef HAVE_HDF5
-//     image_p = new casa::HDF5Image<double> (tile,
+//     image_p = new casa::HDF5Image<float> (tile,
 // 					   csys,
 // 					   filename_p);
 // #else
-    image_p = new casa::PagedImage<double> (tile,
+    image_p = new casa::PagedImage<float> (tile,
 					    csys,
 					    filename_p);
 // #endif
